pyproject.toml
[tool.poetry]
name = "graph-trainstation-test-data"
version = "0.1.0"
description = ""
authors = ["fabioklr <fabiokeller@bluewin.ch>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain-core = "^0.2.24"
langgraph = "^0.1.16"
langchain-openai = "^0.1.19"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.10"
langchainhub = "^0.1.14"
fastapi = "^0.110.0"
langserve = "^0.2.2"
uvicorn = "^0.27.1"
pydantic = "2.8.2"
neo4j = "^5.23.1"
pandas = "^2.2.2"
numpy = "1.26.4"
langchain-anthropic = "^0.1.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
